格式调整

+++ marry.biz/src/main/java/com/tqmall/model/PrimaryKeyInterface.java
@@ -4,5 +4,5 @@ package com.tqmall.model;
  * Created by Minutch on 2015-05-31.
  */
 public interface PrimaryKeyInterface {
-    Long getPrimaryKey();
+    Integer getPrimaryKey();
 }
